+++ src/index.html
@@ -13,7 +13,7 @@
 			<button id="closeButton">X</button>
 
 			<section>
-				<h2>File Import</h2>
+				<h2>File</h2>
 				<button id="ndsFileImportButton">Import .nds</button>
 				<button id="exportButton" disabled>Export</button>
